backend/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
psycopg2-binary==2.9.1
Pillow
django-cors-headers==3.10.0
djangorestframework-simplejwt==5.0.0